--- a/src/jdk.charsets/share/classes/sun/nio/cs/ext/ISO2022_CN_CNS.java
+++ b/src/jdk.charsets/share/classes/sun/nio/cs/ext/ISO2022_CN_CNS.java
@@ -66,18 +66,22 @@ public class ISO2022_CN_CNS extends ISO2022 implements HistoricallyNamedCharset
 
         private static final Charset cns = new EUC_TW();
 
+        private static final byte[] SOD = new byte[] {'$', ')', 'G' };
+        private static final byte[] SS2D = new byte[] {'$', '*', 'H' };
+        private static final byte[] SS3D = new byte[] {'$', '+', 'I' };
+
         public Encoder(Charset cs)
         {
             super(cs);
-            SODesig =  new byte[] {'$', ')', 'G' };
-            SS2Desig = new byte[] {'$', '*', 'H' };
-            SS3Desig = new byte[] {'$', '+', 'I' };
+            SODesig = SOD;
+            SS2Desig = SS2D;
+            SS3Desig = SS3D;
             ISOEncoder = cns.newEncoder();
         }
 
-        private byte[] bb = new byte[4];
+        private final byte[] bb = new byte[4];
         public boolean canEncode(char c) {
-            int n = 0;
+            int n;
             return (c <= '\u007f' ||
                     (n = ((EUC_TW.Encoder)ISOEncoder).toEUC(c, bb)) == 2 ||
                     (n == 4 && bb[0] == SS2 &&