--- a/src/jdk.charsets/share/classes/sun/nio/cs/ext/ISO2022_CN_GB.java
+++ b/src/jdk.charsets/share/classes/sun/nio/cs/ext/ISO2022_CN_GB.java
@@ -29,11 +29,8 @@
 package sun.nio.cs.ext;
 
 import java.nio.charset.Charset;
-import java.nio.ByteBuffer;
-import java.nio.CharBuffer;
 import java.nio.charset.CharsetDecoder;
 import java.nio.charset.CharsetEncoder;
-import java.nio.charset.CoderResult;
 import sun.nio.cs.*;
 import sun.nio.cs.HistoricallyNamedCharset;