@cl4es 2021-02-09
pre-JDK-8259842:
```
Benchmark                                                                    (charsetName)  Mode  Cnt     Score     Error   Units
StringDecode.WithCharset.decodeCharsetName                                   x-ISO-2022-CN-GB  avgt    5  756.293 ±  56.082   ns/op
StringDecode.WithCharset.decodeCharsetName:·gc.alloc.rate.norm               x-ISO-2022-CN-GB  avgt    5  488.040 ±   0.013    B/op
StringDecode.WithCharset.decodeCharsetName                                     ISO-2022-KR          avgt    5  1076.745 ± 299.789   ns/op
StringDecode.WithCharset.decodeCharsetName:·gc.alloc.rate.norm                 ISO-2022-KR         avgt    5   808.068 ±   0.022    B/op
```
Current:
```
Benchmark                                                                    (charsetName)  Mode  Cnt     Score     Error   Units
StringDecode.WithCharset.decodeCharsetName                                   x-ISO-2022-CN-GB  avgt    5  906.281 ± 109.035   ns/op
StringDecode.WithCharset.decodeCharsetName:·gc.alloc.rate.norm               x-ISO-2022-CN-GB  avgt    5  920.077 ±   0.025    B/op
StringDecode.WithCharset.decodeCharsetName                                     ISO-2022-KR          avgt    5  1175.228 ± 144.509   ns/op
StringDecode.WithCharset.decodeCharsetName:·gc.alloc.rate.norm                 ISO-2022-KR         avgt    5  1256.105 ±   0.043    B/op
```
Patched:
```
Benchmark                                                                    (charsetName)  Mode  Cnt    Score     Error   Units
StringDecode.WithCharset.decodeCharsetName                                   x-ISO-2022-CN-GB  avgt    5  783.385 ± 255.048   ns/op
StringDecode.WithCharset.decodeCharsetName:·gc.alloc.rate.norm               x-ISO-2022-CN-GB  avgt    5  584.050 ±   0.016    B/op
StringDecode.WithCharset.decodeCharsetName                                     ISO-2022-KR         avgt    5  814.654 ± 196.860   ns/op
StringDecode.WithCharset.decodeCharsetName:·gc.alloc.rate.norm                 ISO-2022-KR         avgt    5  584.050 ±   0.024    B/op
```